requirements.txt
autopep8==1.5
Click==7.0
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.3
numpy==1.18.1
pyasn1==0.4.8
pycodestyle==2.5.0
PyMySQL==0.9.3
pyparsing==2.4.6
PyQt5==5.13.0
PyQt5-sip==12.7.1
pyqt5-tools==5.13.0.1.5
python-dateutil==2.8.1
python-dotenv==0.11.0
rsa==4.0
six==1.14.0
